+++ terminal.php
@@ -144,6 +144,7 @@ switch ($cmd) {
     case 'ls': case 'cd': case 'mkdir': case 'rmdir': case 'touch': case 'rm': case 'cat':
     case 'wc': case 'more': case 'less': case 'grep': case 'cp': case 'mv':
     case 'head': case 'tail': case 'du': case 'chmod': case 'chown': case 'diff': case 'sort':
+    case 'uniq':
         require __DIR__ . '/commands/filesystem.php';
         break;